Refactoring the whole project

+++ src/main/java/com/example/gates/Repositories/NewsRepository.java
@@ -3,5 +3,6 @@ package com.example.gates.Repositories;
 import com.example.gates.Models.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<News, Long> {
+public interface NewsRepository extends JpaRepository<News,Integer>
+{
 }
